+++ backend/libs/shared/provider/keystore-persistence/src/lib/shared-keystore-persistence.module.ts
@@ -2,6 +2,7 @@ import { Module } from '@nestjs/common';
 import { MongooseModule } from '@nestjs/mongoose';
 import { KeystoreMongoDbService } from './keystore-mongodb.service';
 import { Keystore, KeystoreSchema } from './schema/keystore.schema';
+import { ToggleSeedService } from './toggle-seed.service';
 
 @Module({
   imports: [
@@ -10,7 +11,7 @@ import { Keystore, KeystoreSchema } from './schema/keystore.schema';
     ]),
   ],
   controllers: [],
-  providers: [KeystoreMongoDbService],
+  providers: [KeystoreMongoDbService, ToggleSeedService],
   exports: [KeystoreMongoDbService],
 })
 export class SharedKeystorePersistenceModule {}